% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.maxout.R
\name{b.maxout}
\alias{b.maxout}
\title{maxout statistic for resampling}
\usage{
b.maxout(z, index, dims)
}
\arguments{
\item{z}{data}

\item{index}{resampling index}

\item{dims}{dims of x and y}
}
\value{
maxout statistic
}
\description{
maxout statistic for resampling
}
\examples{
\dontrun{
b.maxout(1:100,51:150,c(50,50))
}
}
