% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corboot.R
\name{corboot}
\alias{corboot}
\title{Bootstrap the correlation}
\usage{
corboot(x, B)
}
\arguments{
\item{x}{data}

\item{B}{Number of Bootstrap replications}
}
\value{
estimated standard deviation
}
\description{
Bootstrap the correlation
}
\examples{
\dontrun{
f(faithful,1000)
}
}
