% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplm.R
\name{cplm}
\alias{cplm}
\title{Change point detection in linear model}
\usage{
cplm(X, Y, target, c, changepoint)
}
\arguments{
\item{X}{the data matrix}

\item{Y}{the response vector}

\item{target}{the target FDR level}

\item{c}{tuning parameter on the interval's length}

\item{changepoint}{oracle location of changepoints for reference}
}
\value{
a list with FDR reported
}
\description{
Change point detection in linear model
}
\examples{
\dontrun{
set.seed(123)
n <- 1000
p <- 60
np <- 4
a <- 0.5
AS <- 2
amplitude <- c(AS,-AS)
changepoint <- c(200,400,600,800) 
beta1 <- rep(0,p)
be1 <- sample(p,3)
beta1[be1] <- sample(amplitude, 3, replace = T)
beta2 <- rep(0,p)
be2 <- sample(p,3)
beta2[be2] <- sample(amplitude, 3, replace = T)
beta3 <- rep(0,p)
be3 <- sample(p,3)
beta3[be3] <- sample(amplitude, 3, replace = T)
beta4 <- rep(0,p)
be4 <- sample(p,3)
beta4[be4] <- sample(amplitude, 3, replace = T)
beta5 <- rep(0,p)
be5 <- sample(p,3)
beta5[be5] <- sample(amplitude, 3, replace = T)
beta <- c(beta1,beta2,beta3,beta4,beta5)
mu <- c(rep(0,p))
target <- 0.2
X <- matrix(rnorm(n*p), nrow = n)
X1 <- X
X2 <- X
X3 <- X
X4 <- X
X5 <- X
X2[(1:changepoint[1]),] <- 0
X3[1:changepoint[2],] <- 0
X4[1:changepoint[3],] <- 0
X5[(1:changepoint[4]),] <- 0
Xo <- cbind(X1,X2,X3,X4,X5)
Y <- Xo\%*\%beta+rnorm(n)
changepoint <- c(200,400,600,800) 
target <- 0.2
c <- 1
cplm(X, Y, target, c, changepoint)
}
}
