% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCov.R
\name{dCov}
\alias{dCov}
\title{Calculate Distance Correlation}
\usage{
dCov(x, y)
}
\arguments{
\item{x}{first sample}

\item{y}{second sample}
}
\value{
Distance Correlation
}
\description{
Calculate Distance Correlation
}
\examples{
\dontrun{
dCov(1:10,5:14)
}
}
