% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMABO.R
\name{EMABO}
\alias{EMABO}
\title{Title EM algorithm for Bolld type}
\usage{
EMABO(N, nA, nB, nO, nAB, n)
}
\arguments{
\item{N}{the number of random initial values}

\item{nA}{number of cases A}

\item{nB}{number of cases B}

\item{nO}{number of cases O}

\item{nAB}{number of cases AB}

\item{n}{number of cases}
}
\value{
the MLE of p, q
}
\description{
Title EM algorithm for Bolld type
}
\examples{
\dontrun{
EMABO(1,28,24,41,70,163)
}
}
