% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_Inde_test.R
\name{KDE_indep_test}
\alias{KDE_indep_test}
\title{Independence Test of Random Variables based on KDE}
\usage{
KDE_indep_test(data)
}
\arguments{
\item{data}{A two dimension data matrix}
}
\value{
test statistic based on KDE and p-value
}
\description{
Independence Test of Random Variables based on KDE
}
\examples{
\dontrun{
KDE_indep_test(faithful)
}
}
