% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_like.R
\name{log_like}
\alias{log_like}
\title{log-likelihood of incomplete likelihood}
\usage{
log_like(p, q, r, nA, nB, nO, nAB)
}
\arguments{
\item{p}{parameter p}

\item{q}{parameter q}

\item{r}{parameter r}

\item{nA}{number of type A}

\item{nB}{number of type B}

\item{nO}{number of type O}

\item{nAB}{number of type AB}
}
\value{
log-likelihood
}
\description{
log-likelihood of incomplete likelihood
}
\examples{
\dontrun{
log_like(1/3,1/3,1/3,28,24,41,70)
}
}
