% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxout.R
\name{maxout}
\alias{maxout}
\title{Compute max out statistic}
\usage{
maxout(x, y)
}
\arguments{
\item{x}{first sample}

\item{y}{second sample}
}
\value{
max out statistic
}
\description{
Compute max out statistic
}
\examples{
\dontrun{
maxout(1:10,5:14)
}
}
