% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_gauss.R
\name{mix_gauss}
\alias{mix_gauss}
\title{Generate the random numbers from the mixture distribution}
\usage{
mix_gauss(n, p1)
}
\arguments{
\item{n}{The sample size}

\item{p1}{Proportion}
}
\value{
Sample following mixture gaussian
}
\description{
Generate the random numbers from the mixture distribution
}
\examples{
\dontrun{
mix_gauss(10,0.4)
}
}
