% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndCov2.R
\name{ndCov2}
\alias{ndCov2}
\title{Title Distance Correlation for Bootstrap}
\usage{
ndCov2(z, ix, dims)
}
\arguments{
\item{z}{all sample}

\item{ix}{index for Bootstrap}

\item{dims}{dimensions of two sample}
}
\value{
test statistic
}
\description{
Title Distance Correlation for Bootstrap
}
\examples{
\dontrun{
ndCov2(1:100,1:100,c(40,60))
}
}
