% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.Metropolis.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{Sampling from laplace distribution}
\usage{
rw.Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{scaling parameter for increment}

\item{x0}{initial value}

\item{N}{sample size}
}
\value{
a list containing sample and number of rejections
}
\description{
Sampling from laplace distribution
}
\examples{
\dontrun{
rw.Metropolis(1,0,100)
}
}
