% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwcpp}
\alias{rwcpp}
\title{Random Walk Metropolis sampler (Laplace Distribution)}
\usage{
rwcpp(sigma, x0, N)
}
\arguments{
\item{sigma}{the scaling parameter for increment distribution}

\item{x0}{the initial value}

\item{N}{the sample size you want}
}
\value{
a random sample of size \code{N}
}
\description{
A MCMC sampler using Rcpp
}
\examples{
\dontrun{
Crw <- rwcpp(1, 0, 2000)
chain <- Crw[[1]]
plot(1:1000,chain[1:1000],type="l",xlab = "Scaling parameter = 1",ylab="Chain",ylim=range(chain))
}
}
