% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu1.R
\name{simu1}
\alias{simu1}
\title{Title Monte Carlo integration with antithetic variables}
\usage{
simu1(n, anti = T)
}
\arguments{
\item{n}{sample size}

\item{anti}{whether to use antithetic variables}
}
\value{
estimated integral
}
\description{
Title Monte Carlo integration with antithetic variables
}
\examples{
\dontrun{
simu1(100,T)
}
}
