% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{A function to calculate skewnwss from samples}
\usage{
skew(x)
}
\arguments{
\item{x}{the sample}
}
\value{
skewness
}
\description{
A function to calculate skewnwss from samples
}
\examples{
\dontrun{
skew(rnorm(100))
}
}
