% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wishart.R
\name{Wishart}
\alias{Wishart}
\title{Title Generate N random samples from Wishart distribution (with degrees of freedom n)}
\usage{
Wishart(N, n, Sigma)
}
\arguments{
\item{N}{sample size}

\item{n}{degree of freedom}

\item{Sigma}{parameter}
}
\value{
N matrix
}
\description{
Title Generate N random samples from Wishart distribution (with degrees of freedom n)
}
\examples{
\dontrun{
Wishart(1,4,diag(2))
}
}
