% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_Inde_test.R
\name{KDE_indep_test}
\alias{KDE_indep_test}
\title{Independence Test of Random Variables based on KDE}
\usage{
KDE_indep_test(data)
}
\arguments{
\item{data}{A two dimension data matrix whose two columns are of interest}
}
\value{
test statistic based on KDE and p-value
}
\description{
Apply KDE method to test the independence. The main idea of this test is to estimate the joint density and marginal density, respectively. And then test if the joint density is the product of two marginal density. The asymptotic distribution of the test statistic is used to compute p-value. The argument data is a matrix or dataframe with two columns. Then the independence of the two columns is tested. It returns the test statistic and p-value.
}
\examples{
\dontrun{
KDE_indep_test(faithful)
}
}
